Rasa/requirements.txt
psycopg2-binary
python-dotenv
rasa
rasa-sdk
SQLAlchemy